Removed unnecessary cast.
git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4039 c5679a2a-0e0e-0410-bba2-8642b794b8c1

+++ Colossus/core/src/main/java/net/sf/colossus/ai/OnTheFlyLegionMove.java
@@ -494,7 +494,7 @@ class OnTheFlyLegionMove implements Collection<LegionMove>
                 {
                     nChoice = length;
                 }
-                parent = byValues.get((int) ((length - nChoice) + rand.nextInt(
+                parent = byValues.get(((length - nChoice) + rand.nextInt(
                         nChoice)));
             }
             return parent;
